+++ packages/cli/uni-builder/tests/default.test.ts
@@ -41,6 +41,60 @@ describe('uni-builder rspack', () => {
 
     process.env.NODE_ENV = NODE_ENV;
   });
+
+  it('should generator rspack config correctly when node', async () => {
+    const { NODE_ENV } = process.env;
+    process.env.NODE_ENV = 'production';
+
+    const rsbuild = await createUniBuilder({
+      bundlerType: 'rspack',
+      config: {
+        environments: {
+          node: {
+            output: {
+              target: 'node',
+            },
+          },
+        },
+      },
+      cwd: '',
+    });
+
+    const {
+      origin: { bundlerConfigs },
+    } = await rsbuild.inspectConfig();
+
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+
+    process.env.NODE_ENV = NODE_ENV;
+  });
+
+  it('should generator rspack config correctly when service-worker', async () => {
+    const { NODE_ENV } = process.env;
+    process.env.NODE_ENV = 'production';
+
+    const rsbuild = await createUniBuilder({
+      bundlerType: 'rspack',
+      config: {
+        environments: {
+          serviceWorker: {
+            output: {
+              target: 'web-worker',
+            },
+          },
+        },
+      },
+      cwd: '',
+    });
+
+    const {
+      origin: { bundlerConfigs },
+    } = await rsbuild.inspectConfig();
+
+    expect(bundlerConfigs[0]).toMatchSnapshot();
+
+    process.env.NODE_ENV = NODE_ENV;
+  });
 });
 
 describe('uni-builder webpack', () => {