+++ packages/devtools/client/src/entries/client/routes/react/state.ts
@@ -1,8 +1,0 @@
-import { $$globals } from '@/entries/client/globals';
-import { WallAgent } from '@/utils/react-devtools';
-
-export const wallAgent = new WallAgent();
-
-$$globals.then(({ mountPoint }) => {
-  wallAgent.bindRemote(mountPoint.remote, 'sendReactDevtoolsData');
-});