+++ packages/runtime/plugin-garfish/src/cli/code.ts
@@ -6,6 +6,7 @@ import type {
 } from '@modern-js/app-tools';
 import { fs } from '@modern-js/utils';
 import { Entrypoint } from '@modern-js/types';
+import type { MaybeAsync } from '@modern-js/core';
 import * as template from './template';
 import { generateAsyncEntryCode } from './utils';
 
@@ -16,6 +17,7 @@ export const generateCode = async (
   entrypoints: Entrypoint[],
   appContext: IAppContext,
   config: NormalizedConfig<AppTools>,
+  appendEntryCode: (input: { entrypoint: Entrypoint }) => MaybeAsync<string[]>,
 ) => {
   const { mountId } = config.html;
   const { enableAsyncEntry } = config.source;
@@ -23,7 +25,9 @@ export const generateCode = async (
     appContext;
   await Promise.all(
     entrypoints.map(async entrypoint => {
-      const { entryName, isAutoMount, entry, customEntry } = entrypoint;
+      const { entryName, isAutoMount, entry, customEntry, customBootstrap } =
+        entrypoint;
+      const appendCode = await appendEntryCode({ entrypoint });
 
       if (isAutoMount) {
         // index.jsx
@@ -34,7 +38,9 @@ export const generateCode = async (
           entry,
           entryName,
           customEntry,
+          customBootstrap,
           mountId,
+          appendCode,
         });
         const indexFile = path.resolve(
           internalDirectory,
@@ -47,7 +53,11 @@ export const generateCode = async (
             internalDirectory,
             `./${entryName}/${ENTRY_BOOTSTRAP_FILE_NAME}`,
           );
-          fs.outputFileSync(bootstrapFile, generateAsyncEntryCode(), 'utf8');
+          fs.outputFileSync(
+            bootstrapFile,
+            generateAsyncEntryCode(appendCode),
+            'utf8',
+          );
         }
       }
     }),