+++ packages/runtime/plugin-runtime/tests/ssr/renderString.test.ts
@@ -2,11 +2,9 @@
  * @jest-environment node
  */
 import path from 'path';
-import { fs } from '@modern-js/utils';
-import {
-  renderString,
-  RenderOptions,
-} from '../../src/core/server/index.server';
+import { fs, createLogger } from '@modern-js/utils';
+import React from 'react';
+import { renderString, RenderOptions } from '../../src/core/server/server';
 import { getInitialContext } from '../../src/core/context/runtime';
 import App from './fixtures/string-ssr/App';
 
@@ -66,6 +64,7 @@ describe('test render', () => {
         routeManifest: {} as any,
       },
       loaderContext: new Map(),
+      logger: createLogger(),
       params: {},
       config: {
         ssr: true,
@@ -73,7 +72,11 @@ describe('test render', () => {
       onTiming,
     };
 
-    const html = await renderString(request as any, App, renderOptions);
+    const serverRoot = React.createElement(App, {
+      _internal_context: { ssr: true },
+    });
+
+    const html = await renderString(request as any, serverRoot, renderOptions);
 
     expect(html).toMatchSnapshot();
     expect(onTiming).toBeCalled();