+++ packages/server/core/src/constants.ts
@@ -17,12 +17,15 @@ export const REPLACE_REG = {
   },
 };
 
-export enum ServerReportTimings {
+export enum ServerTimings {
   SERVER_HANDLE_REQUEST = 'server-handle-request',
   SERVER_MIDDLEWARE = 'server-middleware',
   SERVER_HOOK_AFTER_RENDER = 'server-hook-after-render',
   SERVER_HOOK_AFTER_MATCH = 'server-hook-after-match',
 }
 
+export const SERVER_TIMING = 'Server-Timing';
+
 export const X_RENDER_CACHE = 'x-render-cache';
+
 export const X_MODERNJS_RENDER = 'x-modernjs-render';