+++ tests/integration/devtools/modern.config.ts
@@ -10,7 +10,6 @@ export default defineConfig({
   },
   source: {
     mainEntryName: 'main',
-    // preEntry: ['./src/prelude.css'],
   },
   output: {
     // disable polyfill and ts checker to make test faster