+++ tests/integration/garfish/app-dashboard/src/index.tsx
@@ -1,0 +1,12 @@
+const sleep = () =>
+  new Promise(resolve => {
+    console.log('custom bootstrap');
+    setTimeout(resolve, 300);
+  });
+
+export default (_App: React.ComponentType, bootstrap: () => void) => {
+  // do something before bootstrap...
+  sleep().then(() => {
+    bootstrap();
+  });
+};