+++ packages/cli/plugin-changeset/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @modern-js/plugin-changeset
 
+## 2.55.0
+
+### Patch Changes
+
+- Updated dependencies [bbcf55a]
+  - @modern-js/utils@2.55.0
+  - @modern-js/plugin-i18n@2.55.0
+
 ## 2.54.6
 
 ### Patch Changes