+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/server-utils
 
+## 2.55.0
+
+### Patch Changes
+
+- Updated dependencies [bbcf55a]
+  - @modern-js/utils@2.55.0
+  - @modern-js/babel-compiler@2.55.0
+  - @modern-js/babel-plugin-module-resolver@2.55.0
+
 ## 2.54.6
 
 ### Patch Changes